backend/requirements.txt
google-cloud-bigquery
google-cloud-storage
rdflib
pandas
google-generativeai
hnswlib
numpy
requests
flask